+++ docs/src/index.md
@@ -1,0 +1,6 @@
+# HypergeoMat.jl Documentation
+
+```@autodocs
+Modules = [HypergeoMat]
+Order   = [:type, :function]
+```
