+++ test/runtests.jl
@@ -1,0 +1,4 @@
+using HypergeoMat
+using Test
+
+@test true
